+++ lib/sidekiq/client.rb
@@ -1,20 +1,36 @@
 require 'multi_json'
 require 'redis'
 
+require 'sidekiq/redis_connection'
+require 'sidekiq/middleware/chain'
+require 'sidekiq/middleware/client/resque_web_compatability'
+require 'sidekiq/middleware/client/unique_jobs'
+
 module Sidekiq
   class Client
+    def self.middleware
+      @middleware ||= Middleware::Chain.new
+    end
 
     def self.redis
       @redis ||= begin
-        # autoconfig for Heroku
-        hash = {}
-        hash[:url] = ENV['REDISTOGO_URL'] if ENV['REDISTOGO_URL']
-        Redis.connect(hash)
+        RedisConnection.create
       end
     end
 
     def self.redis=(redis)
       @redis = redis
+    end
+
+    def self.ignore_duplicate_jobs=(value)
+      @ignore_duplicate_jobs = value
+      if @ignore_duplicate_jobs
+        middleware.register do
+          use Middleware::Client::UniqueJobs, Client.redis
+        end
+      else
+        middleware.unregister(Middleware::Client::UniqueJobs)
+      end
     end
 
     # Example usage:
@@ -24,7 +40,9 @@ module Sidekiq
       raise(ArgumentError, "Message must include a class and set of arguments: #{item.inspect}") if !item['class'] || !item['args']
 
       item['class'] = item['class'].to_s if !item['class'].is_a?(String)
-      redis.rpush("queue:#{queue}", MultiJson.encode(item))
+      middleware.invoke(item, queue) do
+        redis.rpush("queue:#{queue}", MultiJson.encode(item))
+      end
     end
 
     # Please use .push if possible instead.