+++ Changes.md
@@ -2,3 +2,11 @@
 -----------
 
 - Initial release!
+
+0.5.2
+-----------
+
+- Refactored middleware support, introducing ability to add client/server middleware (Ryan)
+- Added middleware for ignoring duplicate jobs (Ryan)
+- Added middleware for displaying jobs in resque-web dashboard (Max)
+- Added redis namespacing support (Max)